slices: make slice equality comparator more flexible (#100)
This PR enables SliceContainsFunc to accept a different type as the
target from the slice elements type. For example if we have a slice
of Person objects but want to detect if the slice contains a person
with a particular name (string), we can now use a comparator of
signature func(Person, string) bool - rather than needing a whole
Person struct for the target.